requirements.dev.txt
pytest==6.2.5
pytest-aiohttp==0.3.0
pytest-cov==3.0.0
pytest-mock==3.6.1
